Restructured files again because I didn't want to install react-rewired and the path was wrong for npm start. I upgraded to React 3.4.0 because I was getting an error where path was undefined. Project works now like before.
