+++ src/redux/reducers.js
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+import initialState from './state-old';
+
+const user = ( state = [], action ) => {
+  switch ( action.type ) {
+    case 'LOGIN':
+      return true
+  }
+}
+
